backend/requirements.txt
asgiref==3.8.1
Django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
sqlparse==0.5.3
django-cors-headers
psycopg2-binary
python-dotenv
openai